requirements/deploy_requirements.txt
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
docutils==0.16
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
keyring==21.3.1
mccabe==0.6.1
packaging==20.4
pkginfo==1.5.0.1
psycopg2-binary==2.8.5
PTable==0.9.2
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
python-dotenv==0.14.0
PyYAML==5.3.1
readme-renderer==26.0
-e git+git@github.com:JesseMaitland/red-scope.git@4625d578573145808a9a3bafc483e28137b68af7#egg=redscope
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
sqlparse==0.3.1
tqdm==4.48.2
twine==3.2.0
urllib3==1.25.10
webencodings==0.5.1
zipp==3.1.0
